refactor(cli): stop utils/ importing serve/ and the UI layer (#9146)
packages/cli/src/utils is the directory every other directory imports
(83 from ui, 30 from the package root, 23 from config), so it only
works as a leaf. It currently imports back into six siblings, which is
what makes the package's directory graph cyclic.

Three moves, each decided by who actually consumes the module:

- serve/fast-path-argv.ts moves down into utils/. It is 25 lines with
  no imports, doing pure argv string manipulation, consumed from three
  layers. A utility filed under the daemon; relocate the target rather
  than rework the caller. Renamed serve-fast-path-argv.ts since
  fast-path-argv is ambiguous among general helpers.
- utils/windowTitle.ts moves up into ui/utils/. Its only consumers are
  AppContainer and startInteractiveUI, and it imports StreamingState,
  ICON and the OSC-8 helper.
- utils/systemInfo.ts imports formatMemoryUsage from core directly
  instead of through ui/utils/formatters.ts, a one-line re-export of
  the same symbol. It cannot move into ui/ because the daemon status
  provider consumes it, so the import is inverted instead.

Adds a no-restricted-imports rule for utils/** rejecting serve/. Scoped
to serve/ only: that direction is now completely clean, so the rule
states something true. The remaining ui/, config/, i18n/ and
nonInteractive/ edges are tracked in #9146.

Runtime upward imports out of cli/src/utils: 25 -> 20; utils -> serve
1 -> 0, eliminating that cycle; utils -> ui 7 -> 3.
