+++ packages/cli/src/serve/fast-path.ts
@@ -10,7 +10,7 @@ import {
   isValidMemoryBudgetMb,
   memoryBudgetRangeError,
 } from '@qwen-code/acp-bridge/daemonMemoryBudget';
-import { normalizeServeFastPathArgv } from './fast-path-argv.js';
+import { normalizeServeFastPathArgv } from '../utils/serve-fast-path-argv.js';
 import type { ServeFastPathSettings } from './fast-path-settings.js';
 import { RUNTIME_STARTUP_CANCELLED_MESSAGE } from './runtime-startup-errors.js';
 import type { ServeOptions } from './types.js';
