+++ packages/cli/src/ui/AppContainer.tsx
@@ -158,7 +158,7 @@ import {
   formatSessionWindowTitle,
   titleStatusPrefix,
   writeTerminalTitle,
-} from '../utils/windowTitle.js';
+} from './utils/windowTitle.js';
 import { clearScreen } from '../utils/stdioHelpers.js';
 import { useTextBuffer } from './components/shared/text-buffer.js';
 import { useLogger } from './hooks/useLogger.js';
