+++ packages/cli/src/utils/systemInfo.ts
@@ -13,9 +13,9 @@ import {
   IdeClient,
   AuthType,
   createDebugLogger,
+  formatMemoryUsage,
   type LspStatusSnapshot,
 } from '@qwen-code/qwen-code-core';
-import { formatMemoryUsage } from '../ui/utils/formatters.js';
 import { GIT_COMMIT_INFO } from '../generated/git-commit.js';
 
 const debugLogger = createDebugLogger('STATUS');
